django_app/apps/api/serializers/language.py
from rest_framework import serializers

from apps.core import models


class LanguageSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Language
        fields = '__all__'
