django_app/requirements.txt
django==3.0.6
djangorestframework==3.11.0
psycopg2-binary==2.8.3
django-filter==2.2.0
markdown==3.2.1
gunicorn==20.0.4
